+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/api2/model/dto/MessageDtos.kt
@@ -92,6 +92,7 @@ internal data class DownstreamMessageDto(
     val type: String,
     val updated_at: Date,
     val user: DownstreamUserDto,
+    val moderation_details: DownstreamModerationDetailsDto? = null,
 
     val extraData: Map<String, Any>,
 ) : ExtraDataDto