+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/plugin/listener/internal/SendMessageListenerDatabase.kt
@@ -16,14 +16,12 @@
 
 package io.getstream.chat.android.offline.plugin.listener.internal
 
-import io.getstream.chat.android.client.errors.ChatErrorCode
 import io.getstream.chat.android.client.errors.isPermanent
 import io.getstream.chat.android.client.extensions.enrichWithCid
 import io.getstream.chat.android.client.extensions.internal.users
 import io.getstream.chat.android.client.persistance.repository.MessageRepository
 import io.getstream.chat.android.client.persistance.repository.UserRepository
 import io.getstream.chat.android.client.plugin.listeners.SendMessageListener
-import io.getstream.chat.android.client.utils.internal.toMessageSyncDescription
 import io.getstream.chat.android.models.Message
 import io.getstream.chat.android.models.SyncStatus
 import io.getstream.log.StreamLog
@@ -61,7 +59,7 @@ internal class SendMessageListenerDatabase(
 
         when (result) {
             is Result.Success -> handleSendMessageSuccess(cid, result.value)
-            is Result.Failure -> handleSendMessageFail(message, result.value)
+            is Result.Failure -> handleSendMessageFailure(message, result.value)
         }
     }
 
@@ -70,28 +68,19 @@ internal class SendMessageListenerDatabase(
         processedMessage: Message,
     ) {
         processedMessage.enrichWithCid(cid)
-            .copy(
-                syncStatus = SyncStatus.COMPLETED,
-                syncDescription = null,
-            )
+            .copy(syncStatus = SyncStatus.COMPLETED)
             .also { message ->
                 userRepository.insertUsers(message.users())
                 messageRepository.insertMessage(message)
             }
     }
 
-    private suspend fun handleSendMessageFail(
+    private suspend fun handleSendMessageFailure(
         message: Message,
         error: Error,
     ) {
         val isPermanentError = error.isPermanent()
-        val isMessageModerationFailed = error is Error.NetworkError &&
-            error.serverErrorCode == ChatErrorCode.MESSAGE_MODERATION_FAILED.code
-
-        StreamLog.w(TAG) {
-            "[handleSendMessageFail] isPermanentError: $isPermanentError" +
-                ", isMessageModerationFailed: $isMessageModerationFailed"
-        }
+        StreamLog.w(TAG) { "[handleSendMessageFailure] isPermanentError: $isPermanentError" }
 
         message.copy(
             syncStatus = if (isPermanentError) {
@@ -99,7 +88,6 @@ internal class SendMessageListenerDatabase(
             } else {
                 SyncStatus.SYNC_NEEDED
             },
-            syncDescription = error.toMessageSyncDescription(),
             updatedLocallyAt = Date(),
         ).also { parsedMessage ->
             userRepository.insertUsers(parsedMessage.users())