+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/repository/domain/message/internal/MessageMapper.kt
@@ -17,7 +17,6 @@
 package io.getstream.chat.android.offline.repository.domain.message.internal
 
 import io.getstream.chat.android.models.Message
-import io.getstream.chat.android.models.MessageSyncDescription
 import io.getstream.chat.android.models.Reaction
 import io.getstream.chat.android.models.User
 import io.getstream.chat.android.offline.repository.domain.message.attachment.internal.AttachmentEntity
@@ -54,7 +53,6 @@ internal suspend fun MessageEntity.toModel(
         reactionCounts = reactionCounts.toMutableMap(),
         reactionScores = reactionScores.toMutableMap(),
         syncStatus = syncStatus,
-        syncDescription = buildMessageSyncDescription(),
         shadowed = shadowed,
         i18n = i18n,
         latestReactions = (latestReactions.map { it.toModel(getUser) }).toMutableList(),
@@ -73,6 +71,7 @@ internal suspend fun MessageEntity.toModel(
         pinnedBy = pinnedByUserId?.let { getUser(it) },
         skipEnrichUrl = skipEnrichUrl,
         skipPushNotification = skipPushNotification,
+        moderationDetails = moderationDetails?.toModel(),
     )
 }
 
@@ -84,8 +83,6 @@ internal fun Message.toEntity(): MessageEntity = MessageEntity(
         text = text,
         html = html,
         syncStatus = syncStatus,
-        syncType = syncDescription?.type,
-        syncContent = syncDescription?.content?.toEntity(),
         type = type,
         replyCount = replyCount,
         deletedReplyCount = deletedReplyCount,
@@ -114,6 +111,7 @@ internal fun Message.toEntity(): MessageEntity = MessageEntity(
         pinnedByUserId = pinnedBy?.id,
         skipPushNotification = skipPushNotification,
         skipEnrichUrl = skipEnrichUrl,
+        moderationDetails = moderationDetails?.toEntity(),
     ),
     attachments = attachments.mapIndexed { index, attachment -> attachment.toEntity(id, index) },
     latestReactions = latestReactions.map(Reaction::toEntity),
@@ -158,6 +156,7 @@ internal suspend fun ReplyMessageEntity.toModel(
             pinnedAt = pinnedAt,
             pinExpires = pinExpires,
             pinnedBy = pinnedByUserId?.let { getUser(it) },
+            moderationDetails = moderationDetails?.toModel(),
         )
     }
 }
@@ -171,7 +170,6 @@ internal fun Message.toReplyEntity(): ReplyMessageEntity =
             text = text,
             html = html,
             syncStatus = syncStatus,
-            syncType = syncDescription?.type,
             type = type,
             replyCount = replyCount,
             deletedReplyCount = deletedReplyCount,
@@ -193,16 +191,7 @@ internal fun Message.toReplyEntity(): ReplyMessageEntity =
             pinnedAt = pinnedAt,
             pinExpires = pinExpires,
             pinnedByUserId = pinnedBy?.id,
+            moderationDetails = moderationDetails?.toEntity(),
         ),
         attachments = attachments.mapIndexed { index, attachment -> attachment.toReplyEntity(id, index) },
     )
-
-private fun MessageEntity.buildMessageSyncDescription(): MessageSyncDescription? = with(messageInnerEntity) {
-    if (syncType == null || syncContent == null) {
-        return null
-    }
-    return MessageSyncDescription(
-        syncType,
-        syncContent.toModel(),
-    )
-}